langraph-email-agent/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain>=0.1.0
langchain-openai>=0.0.2
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0